Added cart fragment nested recyclerviews

+++ app/src/main/java/edu/curtin/foodapp/model/orders/Orders.java
@@ -58,10 +58,10 @@ public class Orders {
     }
 
     // Calculate total of all restaurants in this order
-    public float getTotalOrderPrice(){
-        float total = 0;
+    public double getTotalOrderPrice(){
+        double total = 0;
         for (RestaurantOrder order: getOrders()){
-            float restaurantTotal = order.getRestaurantTotal();
+            double restaurantTotal = order.getRestaurantTotal();
             total += restaurantTotal;
         }
         return total;
